Refactor variable names for clarity and consistency
